requirements.txt
# Bot Factory Dependencies
fastapi>=0.100.0
pydantic>=2.0.0
boto3>=1.28.0
anthropic>=0.25.0
numpy>=1.24.0
pyyaml>=6.0
uvicorn>=0.23.0
flake8>=7.0.0
black>=24.0.0
